// assets/script/protobuf/holdem/code_pb.d.ts
export interface CodeMap {
// ...
  MSG_D_MJ_ENTER_ROOM: 3001;
  MSG_D_MJ_SEATED: 3002;
  MSG_D_MJ_BRING_IN: 3003;
  MSG_D_MJ_EXCHANGE_TILES: 3004;
  MSG_D_MJ_VOID_SUIT: 3005;
  MSG_D_MJ_ACTION: 3006;
  MSG_D_MJ_AUTO_OP_ACTIVE: 3007;
  MSG_D_MJ_STANDUP_ACTIVE: 3008;
  MSG_D_MJ_LEAVE: 3009;
  MSG_D_MJ_KEEP_SEAT_ACTIVE: 3010;
  MSG_D_MJ_ADD_TIME: 3011;
  MSG_D_MJ_BROADCAST_MSG: 3012;
  MSG_D_MJ_PRIVATE_MSG: 3013;
  MSG_D_MJ_ROOMERS: 3014;
  MSG_D_MJ_SET_AUTO_ON_TABLE: 3015;
  MSG_D_MJ_OBSERVERS: 3016;
  MSG_D_MJ_SYNC_ENTER: 3017;
  MSG_D_MJ_REPLAY: 3018;
  MSG_D_MJ_CHAT_MEMBERS: 3019;
  MSG_D_MJ_RAISE: 3020;
  MSG_D_MJ_REJECT_BRING_IN: 3021;
  MSG_S_MJ_SEATED_OTHERS: 3051;
  MSG_S_MJ_START_INFO: 3052;
  MSG_S_MJ_EXCHANGE_TILES_COMPLETE: 3053;
  MSG_S_MJ_VOID_SUIT_COMPLETE: 3054;
  MSG_S_MJ_CHIPS_CHANGE: 3055;
  MSG_S_MJ_ACTION_ALL: 3056;
  MSG_S_MJ_AUTO_OP: 3057;
  MSG_S_MJ_STANDUP: 3058;
  MSG_S_MJ_KEEP_SEAT: 3059;
  MSG_S_MJ_RESULT: 3060;
  MSG_S_MJ_ADD_TIME_OTHERS: 3061;
  MSG_S_MJ_LEAVE_NOTIFICATION: 3062;
  MSG_S_MJ_BRING_IN_FAIL: 3063;
  MSG_S_MJ_HAND_CLEAR: 3064;
  MSG_S_MJ_GET_MSG: 3065;
  MSG_S_MJ_WAIT_TURN_ACTION: 3066;
  MSG_S_MJ_ACTION_FAIL: 3067;
  MSG_S_MJ_PREPARE: 3068;
  MSG_S_MJ_RAISE_COMPLETE: 3069;
  MSG_S_MJ_FOLLOW_DEALER: 3070;
  MSG_S_MJ_NEED_BRING_IN: 3071;
  MSG_S_MJ_NEED_BRING_IN_COMPLETE: 3072;
  MSG_S_MJ_VIDEO_MASK_CHANGE: 3073;
  MSG_S_MJ_UP_BLIND: 3074;
  MSG_S_MJ_SYNC_HAND: 3075;
// ...
}

export const Code: CodeMap;